backend/requirements.txt
# ---------------------------------------------------------------------------
# FastAPI and Server
# ---------------------------------------------------------------------------
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ---------------------------------------------------------------------------
# Database
# ---------------------------------------------------------------------------
sqlalchemy>=2.0.35
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
alembic>=1.13.2
greenlet>=3.0.3

# ---------------------------------------------------------------------------
# Authentication and Security
# ---------------------------------------------------------------------------
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# ---------------------------------------------------------------------------
# Validation and Serialization
# ---------------------------------------------------------------------------
pydantic>=2.8.2
pydantic-settings>=2.4.0
email-validator>=2.2.0

# ---------------------------------------------------------------------------
# HTTP Client
# ---------------------------------------------------------------------------
httpx>=0.27.0

# ---------------------------------------------------------------------------
# Testing
# ---------------------------------------------------------------------------
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# ---------------------------------------------------------------------------
# Development
# ---------------------------------------------------------------------------
python-dotenv>=1.0.1
